feat: add logger box and dry-run option

+++ src/lib/run.ts
@@ -6,5 +6,11 @@ import { getUpdatedFileInfo } from "./utils";
 export const run = (optionValues: OptionValues) => {
   const { updatedFileInfo } = getUpdatedFileInfo(optionValues);
 
+  if (optionValues.dryRun) {
+    console.info(updatedFileInfo);
+
+    return;
+  }
+
   fs.writeFileSync(updatedFileInfo.path, updatedFileInfo.content.next);
 };
